[conflict-hand] Conflict hand working  with current system.
I have the conflict hand working with the current system, with validation and such happening on the back end. Cards are able to play despite some cards being played automatically and some need user confirmation.

Unfortunately, this has led to a brittle system that will break easily on changes in the server code.  Also, the code is running slowly. Although, I'm not sure how much is contribution from the computations of ConflictHand or the constant rerendering of the game state.