Refactor Ability Resolver (#2827)
Previously, the ability resolver was a bit of a mish-mash of custom built events and event windows. Due to recent rulings, we have more clarity about how the playing cards/initating abilities process works. This means that the ability resolver now works as follows (which I think is much closer to the way the game rules mandate):

Step 1 & 2 happen before an ability is clicked
After the card is clicked, the ability resolver is added to the pipeline
it snapshots the ability's source
it raises an onCardAbilityInitiated event (any interrupts to ability initiation trigger here)
this adds step 3, 4, 5 and 6 to the pipeline
step 6 raises an onInitiateAbilityEffects event, which checks for cancels
the window for reactions to onCardAbilityInitiated happen next
finally the handler is executed (assuming the ability wasn't cancelled)

* Refactor Ability Resolver

* Remove erroneous edits to CardAbility

* Update abilityresolver.js

Events which are cancelled during targeting shouldn't trigger e.g. WC