aCF - Asako Takahiro (#3590)
* card implementation and test

wrote the card implementation with 2 static effects. one for mil and
one for pol. then I wrote one test to check each effect was working

* fix bug for him counting himself

I made it to where he filters himself out when modifying his skill and add a
test to make sure it is working

* changing to use .getNumberOfParticipants

* used `getNumberOfParticipants`

instead of filtering the cards in play for each player instead just call
getNumberOfParticipants on the game object to make it a lot less LOC